Refactor style lengths per #13584
